Clean up

+++ tests/functional/run.sh
@@ -25,14 +25,16 @@ if [ -z $GOPATH ] ; then
 fi
 
 # Show running machines
-set -x
-_sudo virsh list --all
-free -h
-rpm -qa | grep vagrant
-rpm -qa | grep libvirt
-cat /proc/cpuinfo
-df -h
-exit 0
+if [ 1 -eq 0 ] ; then
+    set -x
+    _sudo virsh list --all
+    free -h
+    rpm -qa | grep vagrant
+    rpm -qa | grep libvirt
+    cat /proc/cpuinfo
+    df -h
+    exit 0
+fi
 
 # Clean up
 rm -f heketi-server > /dev/null 2>&1
